refactor, move formatFilterCombo to its own module
